Move logic into menu
