+++ gatsby/src/components/DetailsPageLayout/DetailsPageLayout.test.tsx
@@ -203,119 +203,4 @@ describe("DetailsPageLayout", () => {
 			expect(screen.getByTestId("body")).toMatchSnapshot();
 		});
 	});
-
-	describe("mobile nav", () => {
-		it("should not render mobile toggle button in static build", () => {
-			const spy = jest
-				.spyOn(React, "useLayoutEffect")
-				.mockImplementation(() => {
-					/*noop*/
-				});
-
-			render(
-				<DetailsPageLayout
-					menu={() => <>Menu</>}
-					titleHtml="Anything"
-					parentBreadcrumbs={[{ href: "/about/", text: "About" }]}
-					{...sections}
-				>
-					Body content
-				</DetailsPageLayout>
-			);
-
-			expect(screen.queryByText("About", { selector: "button" })).toBeNull();
-
-			spy.mockRestore();
-		});
-
-		it("should render mobile anchor to topic menu in static build", () => {
-			const spy = jest
-				.spyOn(React, "useLayoutEffect")
-				.mockImplementation(() => {
-					/*noop*/
-				});
-
-			render(
-				<DetailsPageLayout
-					menu={() => <>Menu</>}
-					titleHtml="Anything"
-					parentBreadcrumbs={[{ href: "/about/", text: "About" }]}
-					{...sections}
-				>
-					Body content
-				</DetailsPageLayout>
-			);
-
-			expect(
-				screen.queryByText("About", {
-					selector: "a.toggleButton",
-				})
-			).toHaveAttribute("href", "#collapsible-menu");
-
-			spy.mockRestore();
-		});
-
-		it("should render collapsed mobile menu button client side", () => {
-			render(
-				<DetailsPageLayout
-					menu={() => <>Menu</>}
-					titleHtml="Anything"
-					parentBreadcrumbs={[{ href: "/about/", text: "About" }]}
-					{...sections}
-				>
-					Body content
-				</DetailsPageLayout>
-			);
-
-			const toggleBtn = screen.getByText("About", {
-				selector: "button",
-			});
-			expect(toggleBtn).toBeInTheDocument();
-			expect(toggleBtn).toHaveAttribute("aria-expanded", "false");
-		});
-
-		it("should render label for screenreaders", () => {
-			render(
-				<DetailsPageLayout
-					menu={() => <>Menu</>}
-					titleHtml="Anything"
-					parentBreadcrumbs={[{ href: "/about/", text: "About" }]}
-					{...sections}
-				>
-					Body content
-				</DetailsPageLayout>
-			);
-
-			const toggleBtn = screen.getByText("About", {
-				selector: "button",
-			});
-			expect(toggleBtn).toBeInTheDocument();
-			expect(toggleBtn).toHaveAttribute("aria-label", "Expand menu for About");
-		});
-
-		it("should collapse toggle button on click", () => {
-			render(
-				<DetailsPageLayout
-					menu={() => <>Menu</>}
-					titleHtml="Anything"
-					parentBreadcrumbs={[{ href: "/about/", text: "About" }]}
-					{...sections}
-				>
-					Body content
-				</DetailsPageLayout>
-			);
-
-			const toggleBtn = screen.getByText("About", {
-				selector: "button",
-			});
-
-			userEvent.click(toggleBtn);
-
-			expect(toggleBtn).toHaveAttribute("aria-expanded", "true");
-			expect(toggleBtn).toHaveAttribute(
-				"aria-label",
-				"Collapse menu for About"
-			);
-		});
-	});
 });
