+++ gatsby/src/components/MedicalDevicesMenu/MedicalDevicesMenu.test.tsx
@@ -36,11 +36,10 @@ describe("MedicalDevicesMenu", () => {
 	});
 
 	it("should highlight current page", () => {
-		(useLocation as jest.Mock).mockImplementationOnce(
-			() =>
-				new URL(
-					"https://bnf-gatsby-tests.nice.org.uk/medical-devices/artificial-saliva-products/"
-				)
+		(useLocation as jest.Mock).mockReturnValue(
+			new URL(
+				"https://bnf-gatsby-tests.nice.org.uk/medical-devices/artificial-saliva-products/"
+			)
 		);
 
 		render(<MedicalDevicesMenu />);
