+++ gatsby/src/components/Menu/Menu.module.scss
@@ -1,0 +1,78 @@
+@use '@nice-digital/nds-core/scss/core';
+
+.toggleButton {
+  appearance: none;
+  background: core.$nds-colour-nice-grey-3;
+  border: 1px solid core.$nds-colour-border;
+  color: core.$nds-colour-link;
+  cursor: pointer;
+  display: block;
+  line-height: 1;
+  /* stylelint-disable indentation, function-parentheses-space-inside */
+  padding: core.rem(
+    core.$nds-spacing-medium core.$nds-spacing-large core.$nds-spacing-medium
+      core.$nds-spacing-small
+  );
+  position: relative;
+  text-align: left;
+  text-decoration: none;
+  width: 100%;
+
+  &:visited {
+    color: core.$nds-colour-link;
+  }
+
+  &:focus {
+    @include core.nds-default-focus-style;
+    z-index: 1;
+  }
+
+  /* stylelint-disable selector-class-pattern */
+  + :global(.stacked-nav) {
+    display: none;
+
+    // Non-JS fallback
+    &:target {
+      display: block;
+    }
+  }
+
+  &[aria-expanded='false'] {
+    margin: 0 0 core.rem(core.$nds-spacing-medium) 0;
+  }
+
+  &[aria-expanded='false'] + :global(.stacked-nav) {
+    display: none;
+  }
+
+  &[aria-expanded='true'] {
+    border-bottom: 0;
+
+    .icon {
+      transform: translateY(-50%) rotate(180deg);
+    }
+
+    + :global(.stacked-nav) {
+      display: block;
+    }
+  }
+
+  @include core.mq($from: sm) {
+    display: none;
+  }
+}
+
+@include core.mq($from: sm) {
+  :global(#collapsible-menu) {
+    display: block;
+  }
+}
+
+.icon {
+  color: core.$nds-colour-link;
+  position: absolute;
+  right: core.rem(core.$nds-spacing-small);
+  top: 50%;
+  transform: translateY(-50%);
+  transition: all 0.25s ease;
+}
