+++ gatsby/src/pages/medical-devices/{BnfMedicalDeviceType.medicalDevice__slug}/{BnfMedicalDeviceType.slug}.tsx
@@ -7,10 +7,10 @@ import { type FeedPrep } from "@nice-digital/gatsby-source-bnf";
 import { Breadcrumbs, Breadcrumb } from "@nice-digital/nds-breadcrumbs";
 import { Grid, GridItem } from "@nice-digital/nds-grid";
 import { PageHeader } from "@nice-digital/nds-page-header";
-import { StackedNav, StackedNavLink } from "@nice-digital/nds-stacked-nav";
 
 import { Layout } from "@/components/Layout/Layout";
 import { MedicalDevicePrepsSection } from "@/components/MedicalDevicePrepsSection/MedicalDevicePrepsSection";
+import { Menu } from "@/components/Menu/Menu";
 import { SEO } from "@/components/SEO/SEO";
 import { useSiteMetadata } from "@/hooks/useSiteMetadata";
 import { QueryResult } from "@/utils";
@@ -40,8 +40,7 @@ const MedicalDeviceTypePage: FC<MedicalDeviceTypePageProps> = ({
 		bnfMedicalDeviceType: { title, medicalDevice, preparations },
 	},
 }) => {
-	const { pathname } = useLocation(),
-		{ siteTitleShort } = useSiteMetadata(),
+	const { siteTitleShort } = useSiteMetadata(),
 		titleNoHtml = striptags(title),
 		hasStackedNav = medicalDevice.medicalDeviceTypes.length > 1,
 		medicalDeviceTitleNoHtml = striptags(medicalDevice.title);
@@ -82,32 +81,20 @@ const MedicalDeviceTypePage: FC<MedicalDeviceTypePageProps> = ({
 			<Grid gutter="loose">
 				{hasStackedNav ? (
 					<GridItem cols={12} md={4} lg={3}>
-						<StackedNav
+						<Menu
 							aria-label={medicalDeviceTitleNoHtml}
 							label={medicalDeviceTitleNoHtml}
 							link={{
 								destination: `/medical-devices/${medicalDevice.slug}/`,
 								elementType: Link,
 							}}
-						>
-							{medicalDevice.medicalDeviceTypes
+							pages={medicalDevice.medicalDeviceTypes
 								.sort((a, b) => a.slug.localeCompare(b.slug))
-								.map((deviceType) => {
-									const pagePath = `/medical-devices/${medicalDevice.slug}/${deviceType.slug}/`;
-									return (
-										<StackedNavLink
-											key={deviceType.title}
-											destination={pagePath}
-											elementType={Link}
-											isCurrent={pathname === pagePath}
-										>
-											<span
-												dangerouslySetInnerHTML={{ __html: deviceType.title }}
-											/>
-										</StackedNavLink>
-									);
-								})}
-						</StackedNav>
+								.map((deviceType) => ({
+									title: deviceType.title,
+									href: `/medical-devices/${medicalDevice.slug}/${deviceType.slug}/`,
+								}))}
+						></Menu>
 					</GridItem>
 				) : null}
 				<GridItem
